+++ CHANGES.md
@@ -1,3 +1,19 @@
+# Version 3.0.0
+- [TASK] Do not register Import module if not luxletter is installed
+- [TASK] Add field pages to registration plugin
+- [TASK] Migrate custom property validation in ActionController
+
+# Version upgrade-typo314-1
+- [FEATURE] Migrate email registration
+- [TASK] Migrate constant editor to site set
+- [TASK] Add field pages to plugins
+- [TASK] Migrate Repository findBy Magic functions
+
+# Version upgrade-typo314-0
+- [TASK] Remove assign to NULL
+- [TASK] Remove RenderCompileStatic from ViewHelpers
+- [WIP] Migrate constant editor to site set
+
 # Version 2.8.0
 - [FEATURE] Add Event after user finished DOI